Changes entity slices to entity types and removes unnecessary code
